+++ cext/cnodefetcher.cpp
