+++ cext/cnodefetcher.hpp
