+++ cext/server.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Rivos Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
-package slurmcprom
+package cext
 
 import (
 	"net/http"
